Create a new crate `syscall` and move `ErrorNo` into it
